# Speed up `test_comparison{,_roi}` for K2IS and FRMS6