These tests are run, if the reference is there, in the default
`pytest tests/` invocation. So they considerably slow down the test runs
for me - mostly because `np.allclose` is quite slow. I see two "fixes"
for this:

1) Don't run the comparison tests by default, mark them "slow" or similar
2) Use a faster comparison function. I would argue that we can require
   equality instead of allowing a tolerance in these cases.
   On one hand, because we are reading the data as integers anyways
   (so it boils down to a `memcmp`), and on the other hand, if we were
   to convert to `float32`, there still shouldn't be a difference,
   because the numbers in range for these formats can easily be
   represented without loss of precision.

I've opted for option 2 for now, because routinely running these tests in development is IMHO valuable.

```
tests/io/datasets/test_k2is.py::test_comparison
before: 1 passed in 107.92s (0:01:47)
after: 1 passed in 14.43s

tests/io/datasets/test_frms6.py::test_comparison
before: 1 passed in 99.43s (0:01:39)
after: 1 passed in 27.13s
```

## Contributor Checklist:

* [N/A] I have added or updated my entry in [the creators.json file](https://github.com/LiberTEM/LiberTEM/blob/master/packaging/creators.json)
* [N/A] I have added [a changelog entry](https://github.com/LiberTEM/LiberTEM/tree/master/docs/source/changelog) for my contribution
* [N/A] I have added/updated documentation for all user-facing changes
* [x] I have added/updated test cases
* [N/A] I have included the [rebuilt production build of the client](https://libertem.github.io/LiberTEM/contributing.html?#building-the-client) (only if changes were made to the GUI)

## Reviewer Checklist:

* [x] `/azp run libertem.libertem-data` passed

<!--

## Please remove this section

Starting by submitting an incomplete pull request (PR) or draft PR is OK. You
can work on the checklist step by step by pushing additional commits into the
PR. Please indicate if you think some items may not be applicable.

You can have a look at [our contributing
docs](https://libertem.github.io/LiberTEM/contributing.html) for more
information on contributing to LiberTEM. Please feel free to ask for
clarification and help, for example in your PR description, with comments or in
our [Gitter channel](https://gitter.im/LiberTEM/Lobby).

Thank you for your contribution!

-->